Add protocol request logic
A request message can be either a subscribe message, or publish message
The message consists of two parts, the header, and the payload.
The header has 4 fields, two of them are implementation details.
The first field is the request type, and the second is the request flags.
The third field is the header checksum, and the forth is the payload checksum.
The payload has two parts, the tags, and the request data.